feature(integration): rename variables and files
