+++ eye-chat/eye-chat/EyeTracking/Alias.swift
@@ -1,5 +1,5 @@
 //
-//  Content.swift
+//  Alias.swift
 //  eye-chat
 //
 //  Created by alexdamascena on 05/04/23.
@@ -7,7 +7,4 @@
 
 import Foundation
 
-enum Content {
-    case onlySections
-    case sectionsAndGrid
-}
+typealias Function = (() -> ())
