+++ eye-chat/eye-chat/EyeTracking/AppState.swift
@@ -6,3 +6,10 @@
 //
 
 import Foundation
+
+
+class AppState: ObservableObject {
+    static let shared = AppState()
+    
+    @Published var gameID = UUID()
+}
