+++ eye-chat/eye-chat/EyeTracking/AttributesTrigger.swift
@@ -1,0 +1,19 @@
+//
+//  AttributesTrigger.swift
+//  eye-chat
+//
+//  Created by alexdamascena on 05/04/23.
+//
+
+import Foundation
+
+
+struct AttributesTrigger {
+    
+    var axisXFirstSectionTrigger: [Function]?
+    var axisXSecondSectionTrigger: [Function]?
+    var upTrigger: Function?
+    var downTrigger: Function?
+    var goBack: Function?
+    
+}
